Refactor compiler to avoid weird swap of contexts
